Separation between auth and users endpoints in API
