+++ server/auth/auth.py
@@ -1,6 +1,5 @@
-from server.models import CustomModel, User
 from server.database import database
-from server.auth.utils import hash_password, verify_password, oauth2_scheme
+from server.auth.utils import verify_password, get_user
 
 import jwt
 from pydantic import BaseModel
@@ -37,15 +36,6 @@ def create_access_token(data: dict):
     encoded_jwt = jwt.encode(to_encode, SECRET_KEY, algorithm=ALGORITHM)
     return encoded_jwt
 
-def get_user(username: str) -> User|None:
-    result = database.execute_read_query(f"SELECT * FROM users WHERE username = ?;", [username])
- 
-    if not result:
-        return None
-
-    user = User.from_database(result[0])
-    return User.model_validate(user)
-
 def update_last_login(username: str) -> None:
     database.execute_query(f"""UPDATE users
                                SET last_login = current_timestamp
@@ -63,105 +53,3 @@ async def login(form_data: Annotated[OAuth2PasswordRequestForm, Depends()]) -> T
 
     access_token = create_access_token({"sub": user.username})
     return Token(access_token=access_token, token_type="bearer")
-
-@router.get("/users/me")
-async def get_current_user(token: str = Depends(oauth2_scheme)) -> User:
-    credentials_exception = HTTPException(
-            status_code=401,
-            headers={"WWW-Authenticate": "Bearer"},
-            detail="Invalid token")
-    try:
-        payload = jwt.decode(token, SECRET_KEY, algorithms=[ALGORITHM])
-        username = payload.get("sub")
-
-        if username == None:
-            raise credentials_exception
-    except jwt.InvalidTokenError:
-        raise credentials_exception
-
-    user = get_user(username)
-
-    if user == None:
-        raise credentials_exception
-
-    return user
-
-@router.get("/users/{username}/details")
-async def get_user_details(username: str, user: User = Depends(get_current_user)) -> User:
-    if user.username != username and not user.is_admin:
-        raise HTTPException(status_code=403, detail="Only admins can get other users' details")
-
-    found_user = get_user(username)
-    if found_user is None:
-        raise HTTPException(status_code=404, detail=f"User '{username}' not found")
-
-    return found_user
-
-@router.get("/users")
-async def get_users(_: User = Depends(get_current_user)) -> list[str]:
-    res = database.execute_read_query("SELECT username FROM users;")
-    usernames = [ entry[0] for entry in res ]
-
-    return usernames
-
-class UserPatch(CustomModel):
-    username: str|None = None
-    password: str|None = None
-    is_admin: bool|None = None
-
-@router.patch("/users/{username}", status_code=200)
-async def update_user(username: str, new_user: UserPatch, user: User = Depends(get_current_user)):
-    if user.username != username and not user.is_admin:
-        raise HTTPException(status_code=403, detail="Only admins can edit other users")
-    if new_user.is_admin and not user.is_admin:
-        raise HTTPException(status_code=403, detail="Only admins can users admin")
-    if new_user.is_admin and username == user.username:
-        raise HTTPException(status_code=403, detail="You may only change the admin status of other users")
-
-    query = "UPDATE users SET "
-
-    values = []
-    for attr in UserPatch.get_attributes():
-        value = getattr(new_user, attr)
-
-        if value == None:
-            continue
-
-        if attr == "password":
-            value = hash_password(value)
-            attr = "password_hash"
-
-        query += f"{attr}=?,"
-        values.append(value)
-
-    if query[-1] == ',':
-        query = query[:-1]
-
-    query += f" WHERE username = ?;"
-    values.append(username)
-    database.execute_query(query, values)
-
-@router.post("/users", status_code=201)
-async def create_user(new_user: UserPatch, user: User = Depends(get_current_user)):
-    if not user.is_admin:
-        raise HTTPException(status_code=403, detail="Only admins change create new users")
-    if not new_user.username or not new_user.password:
-        raise HTTPException(status_code=400, detail="Username and password are required fields")
-
-    if len(new_user.username) < 3:
-        raise HTTPException(status_code=400, detail="Username should be at least 3 characters long")
-
-    password_hash = hash_password(new_user.password)
-    is_admin = new_user.is_admin if new_user.is_admin != None else False
-    database.execute_query(f"INSERT INTO users (username, password_hash, is_admin) VALUES (?, ?, ?)",
-                           [new_user.username, password_hash, is_admin])
-
-@router.delete("/users/{username}", status_code=200)
-async def delete_user(username: str, user: User = Depends(get_current_user)):
-    if not user.is_admin:
-        raise HTTPException(status_code=403, detail="Only admins can delete users")
-    if username == user.username:
-        raise HTTPException(status_code=400, detail="You cannot delete your own user")
-
-    database.execute_query("DELETE FROM flights WHERE username = ?;", [username])
-    database.execute_query("DELETE FROM users WHERE username = ?;", [username])
