+++ server/routers/exporting.py
@@ -1,6 +1,6 @@
-from server.auth.auth import get_current_user
 from server.models import AirportModel, FlightModel, User
 from server.routers.flights import get_flights
+from server.auth.users import get_current_user
 
 from fastapi import APIRouter, Depends
 from fastapi.responses import FileResponse
